Distance of each track point to a given point

Distance of each track point to a given point

dist2point.3d(track, point, groundDistance = FALSE)

Arguments

track

a list containing data.frames with x,y,z coordinates or a data.frame

point

a vector with x, y or x, y, z coordinates

groundDistance

logical: calculate only ground distance in x-y plane?

Value

Returns the distance of each track point to the point.

Examples

dist2point.3d(niclas, c(0, 0, 0))
#>  [1] 2819169 2821469 2823440 2825552 2826797 2831502 2837180 2842031 2845934
#> [10] 2847790 2851064 2855332 2859470 2861141 2862742 2863614 2868419 2873568
#> [19] 2878074 2883691 2886463 2892050 2895537 2901839 2907412 2913614 2918863
#> [28] 2923245 2927654 2929557 2932749