Extract tldCube and autodifference approximation functions

Creates a list consisting of the three dimensional probability distribution cube for turning angle, lift angle and step length (turnLiftStepHist) as well as the uni-dimensional distributions of the differences of the turn angles, lift angles and step lengths with a lag of 1 to maintain minimal level of autocorrelation in each of the terms. Additionally also the distribution of the flight height over the ellipsoid (absolute) and the distribution of flight height over the topography (relative) can be included.

get.densities.3d(turnAngle, liftAngle, stepLength, deltaLift, deltaTurn,
  deltaStep, gradientAngle = NULL, heightEllipsoid = NULL,
  heightTopo = NULL, maxBin = 25)

Arguments

turnAngle
turn angles of the track (t)

liftAngle
lift angles of the track (l)

stepLength
stepLength of the track (d)

deltaLift
auto differences of the turn angles (diff(t))

deltaTurn
auto differences of the lift angles (diff(l))

deltaStep
auto differences of the step length (diff(d))

gradientAngle
NULL or the gardient angles of the track

heightEllipsoid
flight height over the ellipsoid (absolute) or NULL to exclude this distribution

heightTopo
flight height over the topography (relative) or NULL to exclude this distribution

maxBin
numeric scalar, maximum number of bins per dimension of the tld-cube (turnLiftStepHist)

Value

A list containing the tldCube and the autodifferences functions (and additionally the flight height distribution functions)

Examples

niclas <- track.properties.3d(niclas)[2:nrow(niclas),]
P <- get.densities.3d(turnAngle=niclas$t, liftAngle=niclas$l, stepLength=niclas$d,
deltaLift=diff(niclas$t), deltaTurn=diff(niclas$l), deltaStep=diff(niclas$d),
gradientAngle = NULL, heightEllipsoid = NULL, heightTopo = NULL, maxBin = 25)

#>   |TLD cube dimensions: 9 x 14 x 3