Extract tldCube and autodifferences functions from a consistent track

Get densities creates a list consisting of the 3 dimensional probability distribution cube for turning angle, lift angle and step length (turnLiftStepHist) as well as the uni-dimensional distributions of the differences of the turning angles, lift angles and step lengths with a lag of 1 to maintain minimal level of autocorrelation in each of the terms.

get.track.densities.3d(
  track,
  gradientDensity = TRUE,
  heightDistEllipsoid = TRUE,
  DEM = NULL,
  maxBin = 25
)

Arguments

track

a data.frame with 3 columns containing the x,y,z coordinates

gradientDensity

logical: Should a distribution of the gradient angle be extracted and later used in the simulations?

heightDistEllipsoid

logical: Should a distribution of the flight height over ellipsoid be extracted and later used in the sim.cond.3d()?

DEM

a raster containing a digital elevation model, covering the same extent as the track

maxBin

numeric scalar, maximum number of bins per dimension of the tld-cube (turnLiftStepHist)

Value

A list containing the tldCube and the autodifferences functions (and additionally the height distribution function)

Note

The time between the acquisition of fix points of the track must be constant, otherwise this leads to distorted statistic distributions, which increases the probability of dead ends. In this case please check track.split.3d and get.section.densities.3d

Examples

get.track.densities.3d(niclas, heightDist = TRUE)
#>   |TLD cube dimensions: 9 x 14 x 3
#> $tldCube
#> $tldCube$values
#>             turn        lift     step       prob
#> 2   -0.353684014 -0.26142679 1865.097 0.03333333
#> 12   0.170112206 -0.13612579 1865.097 0.03333333
#> 17  -0.179085274 -0.09445602 1865.097 0.03333333
#> 18  -0.004486534 -0.09445602 1865.097 0.03333333
#> 23  -0.353684014 -0.05278625 1865.097 0.03333333
#> 27   0.344710946 -0.05278625 1865.097 0.03333333
#> 32  -0.004486534 -0.01111648 1865.097 0.03333333
#> 33   0.170112206 -0.01111648 1865.097 0.03333333
#> 38  -0.179085274  0.03055329 1865.097 0.03333333
#> 43  -0.878265928  0.07222306 1865.097 0.03333333
#> 81  -0.004486534 -0.09445602 3707.217 0.03333333
#> 95  -0.004486534 -0.01111648 3707.217 0.03333333
#> 105  0.520095381  0.03055329 3707.217 0.03333333
#> 116 -0.004486534  0.11389283 3707.217 0.03333333
#> 119  0.520095381  0.11389283 3707.217 0.03333333
#> 149 -0.353684014 -0.05278625 5549.338 0.03333333
#> 157 -0.179085274 -0.01111648 5549.338 0.10000000
#> 158 -0.004486534 -0.01111648 5549.338 0.06666667
#> 161  0.520095381 -0.01111648 5549.338 0.03333333
#> 164 -0.179085274  0.03055329 5549.338 0.03333333
#> 165 -0.004486534  0.03055329 5549.338 0.06666667
#> 166  0.170112206  0.03055329 5549.338 0.03333333
#> 172 -0.004486534  0.07222306 5549.338 0.03333333
#> 173  0.170112206  0.07222306 5549.338 0.03333333
#> 180  0.170112206  0.11389283 5549.338 0.03333333
#> 189  0.520095381  0.28086360 5549.338 0.03333333
#> 
#> $tldCube$tRes
#> [1] 0.1745987
#> 
#> $tldCube$lRes
#> [1] 0.04166977
#> 
#> $tldCube$dRes
#> [1] 1839.362
#> 
#> 
#> $autoT
#> function (v) 
#> .approxfun(x, y, v, method, yleft, yright, f, na.rm)
#> <bytecode: 0x55b81af7f778>
#> <environment: 0x55b814d6fb30>
#> 
#> $autoL
#> function (v) 
#> .approxfun(x, y, v, method, yleft, yright, f, na.rm)
#> <bytecode: 0x55b81af7f778>
#> <environment: 0x55b814bcc908>
#> 
#> $autoD
#> function (v) 
#> .approxfun(x, y, v, method, yleft, yright, f, na.rm)
#> <bytecode: 0x55b81af7f778>
#> <environment: 0x55b814b86660>
#> 
#> $gDens
#> function (v) 
#> .approxfun(x, y, v, method, yleft, yright, f, na.rm)
#> <bytecode: 0x55b81af7f778>
#> <environment: 0x55b814a8bf10>
#> 
#> $hDistEllipsoid
#> function (v) 
#> .approxfun(x, y, v, method, yleft, yright, f, na.rm)
#> <bytecode: 0x55b81af7f778>
#> <environment: 0x55b814a5a648>
#> 
#> $hDistTopo
#> function (x) 
#> {
#>     1
#> }
#> <bytecode: 0x55b81adf13c8>
#> <environment: 0x55b814fb3320>
#>