Simulates multiple 'gliding & soaring' tracks with a given number of gliding steps

Creates conditional empirical random walks in gliding mode, between a start and end point. The walk is performed on a MODE layer and, if provided, additionally on a background and digital elevation layer. The gliding is simulated with sim.cond.3d and soaring with sim.uncond.3d, therefore soaring is not restricted towards the target and can happen completly free as long as there are good thermal conditions. It is important to extract for every mode in the MODE raster layer a corresponding densities object with get.densities.3d and pass them to the function.

n.sim.glidingSoaring.3d(n.sim = 1, parallel = FALSE, MODE, dGliding,
  dSoaring, qGliding, start = c(0, 0, 0), end = start, a0, g0,
  error = TRUE, smoothTransition = TRUE, glideRatio = 20,
  DEM = NULL, BG = NULL, verbose = FALSE)

Arguments

n.sim
number of simulations to produce

parallel
logical: run computations in parallel (n-1 cores)? Or numeric: the number of nodes (maximum: n - 1 cores)

MODE
raster layer containing the number/index of the mode, which should be used at each location

dGliding
density object returned by the get.densities.3d function for gliding mode

dSoaring
density object returned by the get.densities.3d function for soaring mode

qGliding
the Q probabilites for the steps in gliding mode (qProb.3d)

start
numeric vector of length 3 with the coordinates of the start point

end
numeric vector of length 3 with the coordinates of the end point

a0
initial incoming heading in radian

g0
initial incoming gradient/polar angle in radian

error
logical: add random noise to the turn angle, lift angle and step length to account for errors measurements?

smoothTransition
logical: should the transitions between soaring and the following gliding sections be smoothed? Recommended to avoid dead ends

glideRatio
ratio between vertical and horizontal movement, by default set to 15 meters forward movement per meter vertical movement

DEM
raster layer containing a digital elevation model, covering the area between start and end point

BG
a background raster layer that can be used to inform the choice of steps

verbose
logical: print current mode used?

Value

A list containing 'soaring-gliding' trajectories or NULLs if dead ends have been encountered.

Note

The MODE raster layer must be in the following structure: Gliding pixels have the value 1 and soaring pixel the values 2. NA's are not allowed in the raster.

Examples

# \donttest{
n.sim.glidingSoaring.3d(locsVec, start = c(0,0,0), end=start, a0, g0, dList, qList, MODE)

#> Error in n.sim.glidingSoaring.3d(locsVec, start = c(0, 0, 0), end = start,     a0, g0, dList, qList, MODE): Objekt 'locsVec' nicht gefunden

# }