Q probabilities for n steps

Calculates the Q probability, representing the pull to the target. The number of steps on which the Q prob will be quantified is number of total segments less than one (the last step is defined by the target itself).

qProb.3d(sim, n.locs, parallel = FALSE, maxBin = 25)

Arguments

sim

the result of sim.uncond.3d, or a data frame with at least x,y,z-coordinates, the arrival azimuth and the arrival gradient.

n.locs

number of total segments to be modeled, the length of the desired conditional empirical random walk

parallel

logical: run computations in parallel (n-1 cores)? Or numeric: the number of nodes (maximum: n - 1 cores)

maxBin

numeric scalar, maximum number of bins per dimension of the tld-cube (turnLiftStepHist)

Value

A list containing the Q - tldCubes for every step

Examples

qProb.3d(niclas, 3)
#>   |Extracting Q probabilities for 3 steps
#>   |                                                                              |                                                                      |   0%  |                                                                              |======================================================================| 100%
#>   |Elapsed time: 0s
#> [[1]]
#> [[1]]$values
#>             turn        lift     step       prob
#> 2   -0.353684014 -0.26142679 1865.097 0.03333333
#> 12   0.170112206 -0.13612579 1865.097 0.03333333
#> 17  -0.179085274 -0.09445602 1865.097 0.03333333
#> 18  -0.004486534 -0.09445602 1865.097 0.03333333
#> 23  -0.353684014 -0.05278625 1865.097 0.03333333
#> 27   0.344710946 -0.05278625 1865.097 0.03333333
#> 32  -0.004486534 -0.01111648 1865.097 0.03333333
#> 33   0.170112206 -0.01111648 1865.097 0.03333333
#> 38  -0.179085274  0.03055329 1865.097 0.03333333
#> 43  -0.878265928  0.07222306 1865.097 0.03333333
#> 81  -0.004486534 -0.09445602 3707.217 0.03333333
#> 95  -0.004486534 -0.01111648 3707.217 0.03333333
#> 105  0.520095381  0.03055329 3707.217 0.03333333
#> 116 -0.004486534  0.11389283 3707.217 0.03333333
#> 119  0.520095381  0.11389283 3707.217 0.03333333
#> 149 -0.353684014 -0.05278625 5549.338 0.03333333
#> 157 -0.179085274 -0.01111648 5549.338 0.10000000
#> 158 -0.004486534 -0.01111648 5549.338 0.06666667
#> 161  0.520095381 -0.01111648 5549.338 0.03333333
#> 164 -0.179085274  0.03055329 5549.338 0.03333333
#> 165 -0.004486534  0.03055329 5549.338 0.06666667
#> 166  0.170112206  0.03055329 5549.338 0.03333333
#> 172 -0.004486534  0.07222306 5549.338 0.03333333
#> 173  0.170112206  0.07222306 5549.338 0.03333333
#> 180  0.170112206  0.11389283 5549.338 0.03333333
#> 189  0.520095381  0.28086360 5549.338 0.03333333
#> 
#> [[1]]$tRes
#> [1] 0.1745987
#> 
#> [[1]]$lRes
#> [1] 0.04166977
#> 
#> [[1]]$dRes
#> [1] 1839.362
#> 
#>