Reproduce a track with the eRTG3D

Simulates n tracks with the geometrical properties of the original track, between the same start and end point.

reproduce.track.3d(
  track,
  n.sim = 1,
  parallel = FALSE,
  error = TRUE,
  DEM = NULL,
  BG = NULL,
  filterDeadEnds = TRUE,
  plot2d = FALSE,
  plot3d = FALSE,
  maxBin = 25,
  gradientDensity = TRUE
)

Arguments

track
data.frame with x,y,z coordinates of the original track

n.sim
number of simulations that should be done

parallel
logical: run computations in parallel (n-1 cores)? Or numeric: the number of nodes (maximum: n - 1 cores)

error
logical: add error term to movement in simulation?

DEM
a raster containing a digital elevation model, covering the same extent as the track

BG
a raster influencing the probabilities.

filterDeadEnds
logical: Remove tracks that ended in a dead end?

plot2d
logical: plot tracks on 2-D plane?

plot3d
logical: plot tracks in 3-D?

maxBin
numeric scalar, maximum number of bins per dimension of the tld-cube (turnLiftStepHist)

gradientDensity
logical: Should a distribution of the gradient angle be extracted and used in the simulations (get.densities.3d)?

Value

A list or data.frame containing the simulated track(s) (CERW).

Examples

reproduce.track.3d(niclas[1:10, ])

#>   |TLD cube dimensions: 8 x 15 x 2

#>   |Simulate UERW with 15000 steps

#>   |                                                                              |                                                                      |   0%  |                                                                              |====                                                                  |   5%  |                                                                              |=======                                                               |  10%  |                                                                              |===========                                                           |  15%  |                                                                              |==============                                                        |  20%  |                                                                              |==================                                                    |  25%  |                                                                              |=====================                                                 |  30%  |                                                                              |=========================                                             |  35%  |                                                                              |============================                                          |  40%  |                                                                              |================================                                      |  45%  |                                                                              |===================================                                   |  50%  |                                                                              |=======================================                               |  55%  |                                                                              |==========================================                            |  60%  |                                                                              |==============================================                        |  65%  |                                                                              |=================================================                     |  70%  |                                                                              |=====================================================                 |  75%  |                                                                              |========================================================              |  80%  |                                                                              |============================================================          |  85%  |                                                                              |===============================================================       |  90%  |                                                                              |===================================================================   |  95%  |                                                                              |======================================================================| 100%

#>   |Elapsed time: 0.7s

#>   |Extracting Q probabilities for 10 steps

#>   |                                                                              |                                                                      |   0%  |                                                                              |=========                                                             |  12%  |                                                                              |==================                                                    |  25%  |                                                                              |==========================                                            |  38%  |                                                                              |===================================                                   |  50%  |                                                                              |============================================                          |  62%  |                                                                              |====================================================                  |  75%  |                                                                              |=============================================================         |  88%  |                                                                              |======================================================================| 100%

#>   |Elapsed time: 0.2s

#>   |Simulate CERW with 10 steps

#>   |                                                                              |                                                                      |   0%  |                                                                              |=========                                                             |  12%  |                                                                              |==================                                                    |  25%  |                                                                              |==========================                                            |  38%  |                                                                              |===================================                                   |  50%  |                                                                              |============================================                          |  62%  |                                                                              |====================================================                  |  75%  |                                                                              |=============================================================         |  88%  |                                                                              |======================================================================| 100%

#>   |Elapsed time: 0.2s

#>          x       y        z         a        g           t           l        d
#> 1  2556476 1188336 1283.736 0.4477398 1.531790  0.03391427  0.01041637       NA
#> 2  2558107 1189033 1330.000 0.4037400 1.544719 -0.04159511  0.01041637 1837.581
#> 3  2559792 1189785 1345.066 0.4198537 1.562632  0.03391427  0.01041637 1837.581
#> 4  2561934 1190637 1354.898 0.3785199 1.566530 -0.04159511  0.01041637 2371.273
#> 5  2566452 1192503 1368.356 0.3916651 1.568043 -0.00384042  0.01041637 4772.891
#> 6  2568521 1193597 1347.717 0.4865263 1.579616  0.10942366  0.01041637 2371.273
#> 7  2570498 1194802 1334.384 0.5473249 1.576555  0.07166897 -0.01284079 2371.273
#> 8  2574264 1197085 1298.262 0.5450165 1.578998 -0.00384042  0.01041637 4506.044
#> 9  2576502 1198570 1334.761 0.5858449 1.557205  0.03391427 -0.01284079 2638.120
#> 10 2581749 1201865 1513.318        NA       NA          NA          NA 6198.513
#>               p
#> 1            NA
#> 2  7.434533e-18
#> 3  1.385153e-17
#> 4  5.801531e-18
#> 5  6.022572e-18
#> 6  4.870566e-18
#> 7  1.205326e-17
#> 8  4.654459e-17
#> 9  1.715756e-17
#> 10           NA