Unconditional Empirical Random Walk (UERW) in 3-D

This function creates unconditional walks with prescribed empirical properties (turning angle, lift angle and step length and the auto-differences of them. It can be used for uncon- ditional walks or to seed the conditional walks with comparably long simulations. The conditional walk connecting a given start with a certain end point by a given number of steps needs an attraction term (the Q probability, see qProb.3d) to ensure that the target is approached and hit. In order to calculate the Q probability for each step the distribution of turns and lifts to target and the distribution of distance to target has to be known. They can be derived from the empirical data (ideally), or estimated from an unconditional process with the same properties. Creates a unconditional empirical random walk, with a specific starting point, geometrically similar to the initial trajectory.

sim.uncond.3d(n.locs, start = c(0, 0, 0), a0, g0, densities, error = TRUE)

Arguments

n.locs

the number of locations for the simulated track

start

vector indicating the start point c(x,y,z)

a0

initial heading in radian

g0

initial gradient/polar angle in radian

densities

list object returned by the get.densities.3d function

error

logical: add random noise to the turn angle, lift angle and step length to account for errors measurements?

Value

A 3 dimensional trajectory in the form of a data.frame

Note

Simulations connecting start and end points with more steps than 1/10th or more of the number of steps of the empirical data should rather rely on simulated unconditional walks with the same properties than on the empirical data (factor = 1500).

Random initial heading

For a random initial heading a0 use: sample(atan2(diff(coordinates(track)[,2]), diff(coordinates(track)[,1])),1)

Examples

sim.uncond.3d(
  10, start = c(0, 0, 0), a0 = pi / 2, g0 = pi / 2,
  densities = get.track.densities.3d(niclas)
)
#>   |Simulate UERW with 10 steps
#>   |                                                                              |                                                                      |   0%
#>   |TLD cube dimensions: 9 x 14 x 3
#>   |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#>   |Elapsed time: 0s
#>             x         y          z         a        g           t            l
#> 1      0.0000     0.000    0.00000 1.5707963 1.570796 -0.87826593  0.072223062
#> 2    285.9731  2020.666  -23.05371 1.4302058 1.582092 -0.14059054  0.011295919
#> 3    941.5793  8447.182  -14.10271 1.4691322 1.569411  0.03892644 -0.012681549
#> 4   1611.6803 10986.990  -76.97013 1.3128356 1.594726 -0.15629665  0.025314862
#> 5   4384.4772 15142.982  -20.85036 0.9824380 1.559564 -0.33039756 -0.035161553
#> 6   7169.0484 19188.063 -216.43285 0.9679086 1.610602 -0.01452945  0.051037835
#> 7  11050.6036 23732.014 -397.48702 0.8638548 1.601083 -0.10405380 -0.009518486
#> 8  14826.1815 27382.776 -471.57509 0.7685925 1.584902 -0.09526221 -0.016181208
#> 9  20092.9978 31036.682 -505.13891 0.6065256 1.576032 -0.16206692 -0.008869848
#> 10 22786.4755 31521.777 -416.98242 0.1781899 1.538596 -0.42833573 -0.037436218
#>           d            p
#> 1  1865.097           NA
#> 2  2040.932 1.460992e-04
#> 3  6459.877 2.210422e-04
#> 4  2627.473 3.695195e-05
#> 5  4996.381 6.765565e-05
#> 6  4914.750 1.945267e-04
#> 7  5978.858 4.038032e-04
#> 8  5252.479 4.823305e-04
#> 9  6410.266 4.814161e-04
#> 10 2738.231 2.793233e-05