This function splits the by outliers in the time lag.

The length of timeLag must be the the track's length minus 1 and represents the time passed between the fix point acquisition

track.split.3d(track, timeLag, lag = NULL, tolerance = NULL)

Arguments

track
track data.frame with x, y and z coordinates

timeLag
a numeric vector with the time passed between the fix point acquisition

lag
NULL or a manually chosen lag

tolerance
NULL or a manually chosen tolerance

Value

A list containing the splitted tracks.

Examples

track.split.3d(niclas, timeLag=rep(1, nrow(niclas)-1) + rnorm(nrow(niclas)-1, mean = 0, sd = 0.25))
#>   |Mean time lag: 0.95, tolerance: 0.15, number of splits: 19, proposed change in steps: 0
#> $`0`
#>         x       y        z
#> 1 2556476 1188336 1283.736
#> 2 2558565 1189297 1359.580
#> 
#> $`1`
#>         x       y        z
#> 3 2560477 1189861 1369.713
#> 
#> $`2`
#>         x       y        z
#> 4 2562431 1190668 1391.752
#> 
#> $`3`
#>         x       y        z
#> 5 2563809 1190657 1806.566
#> 6 2567940 1192930 1764.062
#> 7 2572877 1195780 1532.481
#> 
#> $`4`
#>         x       y        z
#> 8 2577051 1198311 1407.334
#> 
#> $`5`
#>         x       y        z
#> 9 2580338 1200499 1282.171
#> 
#> $`6`
#>          x       y        z
#> 10 2581749 1201865 1513.318
#> 
#> $`7`
#>          x       y        z
#> 11 2584181 1204396 1553.903
#> 
#> $`8`
#>          x       y        z
#> 12 2587086 1208264 1355.825
#> 13 2589942 1211928 1180.242
#> 
#> $`9`
#>          x       y        z
#> 14 2591298 1212971 1249.781
#> 15 2592320 1214565 1427.830
#> 
#> $`10`
#>          x       y        z
#> 16 2593267 1214599 1448.141
#> 
#> $`11`
#>          x       y        z
#> 17 2597195 1217541 1565.054
#> 
#> $`12`
#>          x       y        z
#> 18 2601943 1219542 1486.655
#> 19 2606612 1220197 1428.523
#> 20 2611936 1222073 1245.329
#> 
#> $`13`
#>          x       y        z
#> 21 2614472 1223194 1374.984
#> 
#> $`14`
#>          x       y        z
#> 22 2619707 1225189 1299.593
#> 23 2623038 1226295 1637.258
#> 24 2628725 1229013 1449.789
#> 25 2634173 1230518 1380.690
#> 
#> $`15`
#>          x       y        z
#> 26 2640336 1231979 1180.499
#> 
#> $`16`
#>          x       y        z
#> 27 2646688 1230772 1280.248
#> 
#> $`17`
#>          x       y        z
#> 28 2652460 1228745 1369.771
#> 29 2656826 1229809 1330.951
#> 
#> $`18`
#>          x       y       z
#> 30 2659108 1229409 1388.27
#> 31 2662223 1230279 1088.63
#>