Extract tldCube and autodifferences functions from track sections

Creates a list consisting of the 3 dimensional probability distribution cube for turning angle, lift angle and step length (turnLiftStepHist) as well as the uni-dimensional distributions of the differences of the turning angles, lift angles and step lengths with a lag of 1 to maintain minimal level of autocorrelation in each of the terms.

get.section.densities.3d(
  trackSections,
  gradientDensity = TRUE,
  heightDistEllipsoid = TRUE,
  DEM = NULL,
  maxBin = 25
)

Arguments

trackSections
list of track sections got by the track.split.3d function

gradientDensity
logical: Should a distribution of the gradient angle be extracted and later used in the simulations?

heightDistEllipsoid
logical: Should a distribution of the flight height over ellipsoid be extracted and later used in the sim.cond.3d()?

DEM
a raster containing a digital elevation model, covering the same extent as the track sections

maxBin
numeric scalar, maximum number of bins per dimension of the tld-cube (turnLiftStepHist)

Value

A list containing the tldCube and the autodifferences functions (and additionally the height distribution function)

Examples

get.section.densities.3d(list(niclas[1:10, ], niclas[11:nrow(niclas), ]))

#>   |TLD cube dimensions: 7 x 14 x 3

#> $tldCube
#> $tldCube$values
#>            turn        lift     step       prob
#> 2   -0.40356937 -0.26142679 1865.097 0.03448276
#> 11   0.26988291 -0.13612579 1865.097 0.03448276
#> 15  -0.17908527 -0.09445602 1865.097 0.03448276
#> 16   0.04539882 -0.09445602 1865.097 0.03448276
#> 20  -0.40356937 -0.05278625 1865.097 0.03448276
#> 23   0.26988291 -0.05278625 1865.097 0.03448276
#> 28   0.04539882 -0.01111648 1865.097 0.06896552
#> 33  -0.17908527  0.03055329 1865.097 0.03448276
#> 37  -0.85332325  0.07222306 1865.097 0.03448276
#> 70   0.04539882 -0.09445602 3707.217 0.03448276
#> 82   0.04539882 -0.01111648 3707.217 0.03448276
#> 90   0.49515270  0.03055329 3707.217 0.03448276
#> 102  0.49515270  0.11389283 3707.217 0.03448276
#> 128 -0.40356937 -0.05278625 5549.338 0.03448276
#> 135 -0.17908527 -0.01111648 5549.338 0.10344828
#> 136  0.04539882 -0.01111648 5549.338 0.06896552
#> 138  0.49515270 -0.01111648 5549.338 0.03448276
#> 141 -0.17908527  0.03055329 5549.338 0.03448276
#> 142  0.04539882  0.03055329 5549.338 0.06896552
#> 143  0.26988291  0.03055329 5549.338 0.03448276
#> 144  0.49515270  0.03055329 5549.338 0.03448276
#> 148  0.04539882  0.07222306 5549.338 0.03448276
#> 154  0.04539882  0.11389283 5549.338 0.03448276
#> 162  0.49515270  0.28086360 5549.338 0.03448276
#> 
#> $tldCube$tRes
#> [1] 0.2244841
#> 
#> $tldCube$lRes
#> [1] 0.04166977
#> 
#> $tldCube$dRes
#> [1] 1839.362
#> 
#> 
#> $autoT
#> function (v) 
#> .approxfun(x, y, v, method, yleft, yright, f)
#> <bytecode: 0x7f9d45afabe0>
#> <environment: 0x7f9d48bee950>
#> 
#> $autoL
#> function (v) 
#> .approxfun(x, y, v, method, yleft, yright, f)
#> <bytecode: 0x7f9d45afabe0>
#> <environment: 0x7f9d48c0a090>
#> 
#> $autoD
#> function (v) 
#> .approxfun(x, y, v, method, yleft, yright, f)
#> <bytecode: 0x7f9d45afabe0>
#> <environment: 0x7f9d48c25d80>
#> 
#> $gDens
#> function (v) 
#> .approxfun(x, y, v, method, yleft, yright, f)
#> <bytecode: 0x7f9d45afabe0>
#> <environment: 0x7f9d48c410c0>
#> 
#> $hDistEllipsoid
#> function (v) 
#> .approxfun(x, y, v, method, yleft, yright, f)
#> <bytecode: 0x7f9d45afabe0>
#> <environment: 0x7f9d48c63328>
#> 
#> $hDistTopo
#> function (x) 
#> {
#>     1
#> }
#> <bytecode: 0x7f9d4584fea8>
#> <environment: 0x7f9d47a61010>
#>