Reproduce a track with the eRTG3D

Simulates n tracks with the geometrical properties of the original track, between the same start and end point.

reproduce.track.3d(
  track,
  n.sim = 1,
  parallel = FALSE,
  error = TRUE,
  DEM = NULL,
  BG = NULL,
  filterDeadEnds = TRUE,
  plot2d = FALSE,
  plot3d = FALSE,
  maxBin = 25,
  gradientDensity = TRUE
)

Arguments

track

data.frame with x,y,z coordinates of the original track

n.sim

number of simulations that should be done

parallel

logical: run computations in parallel (n-1 cores)? Or numeric: the number of nodes (maximum: n - 1 cores)

error

logical: add error term to movement in simulation?

DEM

a raster containing a digital elevation model, covering the same extent as the track

BG

a raster influencing the probabilities.

filterDeadEnds

logical: Remove tracks that ended in a dead end?

plot2d

logical: plot tracks on 2-D plane?

plot3d

logical: plot tracks in 3-D?

maxBin

numeric scalar, maximum number of bins per dimension of the tld-cube (turnLiftStepHist)

gradientDensity

logical: Should a distribution of the gradient angle be extracted and used in the simulations (get.densities.3d)?

Value

A list or data.frame containing the simulated track(s) (CERW).

Examples

reproduce.track.3d(niclas[1:10, ])
#>   |TLD cube dimensions: 8 x 15 x 2
#>   |Simulate UERW with 15000 steps
#>   |                                                                              |                                                                      |   0%  |                                                                              |====                                                                  |   5%  |                                                                              |=======                                                               |  10%  |                                                                              |===========                                                           |  15%  |                                                                              |==============                                                        |  20%  |                                                                              |==================                                                    |  25%  |                                                                              |=====================                                                 |  30%  |                                                                              |=========================                                             |  35%  |                                                                              |============================                                          |  40%  |                                                                              |================================                                      |  45%  |                                                                              |===================================                                   |  50%  |                                                                              |=======================================                               |  55%  |                                                                              |==========================================                            |  60%  |                                                                              |==============================================                        |  65%  |                                                                              |=================================================                     |  70%  |                                                                              |=====================================================                 |  75%  |                                                                              |========================================================              |  80%  |                                                                              |============================================================          |  85%  |                                                                              |===============================================================       |  90%  |                                                                              |===================================================================   |  95%  |                                                                              |======================================================================| 100%
#>   |Elapsed time: 0.6s
#>   |Extracting Q probabilities for 10 steps
#>   |                                                                              |                                                                      |   0%  |                                                                              |=========                                                             |  12%  |                                                                              |==================                                                    |  25%  |                                                                              |==========================                                            |  38%  |                                                                              |===================================                                   |  50%  |                                                                              |============================================                          |  62%  |                                                                              |====================================================                  |  75%  |                                                                              |=============================================================         |  88%  |                                                                              |======================================================================| 100%
#>   |Elapsed time: 0.2s
#>   |Simulate CERW with 10 steps
#>   |                                                                              |                                                                      |   0%  |                                                                              |=========                                                             |  12%  |                                                                              |==================                                                    |  25%  |                                                                              |==========================                                            |  38%  |                                                                              |===================================                                   |  50%  |                                                                              |============================================                          |  62%  |                                                                              |====================================================                  |  75%  |                                                                              |=============================================================         |  88%  |                                                                              |======================================================================| 100%
#>   |Elapsed time: 0.2s
#>          x       y        z         a        g           t           l        d
#> 1  2556476 1188336 1283.736 0.4477398 1.531790 -0.01459126  0.05666516       NA
#> 2  2559108 1189599 1383.410 0.4474605 1.536666 -0.01459126  0.01005079 2897.952
#> 3  2560541 1190472 1369.192 0.5474283 1.579271  0.09430376  0.03335797 1774.368
#> 4  2561842 1191163 1348.518 0.4875632 1.584828 -0.05088960  0.01005079 1547.517
#> 5  2564511 1192634 1343.542 0.5037858 1.572429  0.02170708 -0.01325640 3122.669
#> 6  2568220 1194564 1401.042 0.4799305 1.557044 -0.01459126 -0.01325640 4246.253
#> 7  2571643 1196289 1401.435 0.4667034 1.570694 -0.01459126  0.01005079 3796.819
#> 8  2573761 1197488 1432.700 0.5152093 1.557955  0.05800542 -0.01325640 2448.519
#> 9  2576706 1199238 1402.039 0.5361255 1.579748  0.02170708  0.03335797 3347.386
#> 10 2581749 1201865 1513.318        NA       NA          NA          NA 5687.082
#>               p
#> 1            NA
#> 2  1.121029e-18
#> 3  1.419986e-18
#> 4  3.330758e-18
#> 5  2.119704e-18
#> 6  1.971857e-17
#> 7  1.474332e-17
#> 8  5.791263e-18
#> 9  5.301822e-18
#> 10           NA