Reproduce a track with the eRTG3D

Simulates n tracks with the geometrical properties of the original track, between the same start and end point.

reproduce.track.3d(
  track,
  n.sim = 1,
  parallel = FALSE,
  error = TRUE,
  DEM = NULL,
  BG = NULL,
  filterDeadEnds = TRUE,
  plot2d = FALSE,
  plot3d = FALSE,
  maxBin = 25,
  gradientDensity = TRUE
)

Arguments

track
data.frame with x,y,z coordinates of the original track

n.sim
number of simulations that should be done

parallel
logical: run computations in parallel (n-1 cores)? Or numeric: the number of nodes (maximum: n - 1 cores)

error
logical: add error term to movement in simulation?

DEM
a raster containing a digital elevation model, covering the same extent as the track

BG
a raster influencing the probabilities.

filterDeadEnds
logical: Remove tracks that ended in a dead end?

plot2d
logical: plot tracks on 2-D plane?

plot3d
logical: plot tracks in 3-D?

maxBin
numeric scalar, maximum number of bins per dimension of the tld-cube (turnLiftStepHist)

gradientDensity
logical: Should a distribution of the gradient angle be extracted and used in the simulations (get.densities.3d)?

Value

A list or data.frame containing the simulated track(s) (CERW).

Examples

reproduce.track.3d(niclas[1:10, ])
#>   |TLD cube dimensions: 8 x 15 x 2
#>   |Simulate UERW with 15000 steps
#>   |                                                                              |                                                                      |   0%  |                                                                              |====                                                                  |   5%  |                                                                              |=======                                                               |  10%  |                                                                              |===========                                                           |  15%  |                                                                              |==============                                                        |  20%  |                                                                              |==================                                                    |  25%  |                                                                              |=====================                                                 |  30%  |                                                                              |=========================                                             |  35%  |                                                                              |============================                                          |  40%  |                                                                              |================================                                      |  45%  |                                                                              |===================================                                   |  50%  |                                                                              |=======================================                               |  55%  |                                                                              |==========================================                            |  60%  |                                                                              |==============================================                        |  65%  |                                                                              |=================================================                     |  70%  |                                                                              |=====================================================                 |  75%  |                                                                              |========================================================              |  80%  |                                                                              |============================================================          |  85%  |                                                                              |===============================================================       |  90%  |                                                                              |===================================================================   |  95%  |                                                                              |======================================================================| 100%
#>   |Elapsed time: 0.9s
#>   |Extracting Q probabilities for 10 steps
#>   |                                                                              |                                                                      |   0%  |                                                                              |=========                                                             |  12%  |                                                                              |==================                                                    |  25%  |                                                                              |==========================                                            |  38%  |                                                                              |===================================                                   |  50%  |                                                                              |============================================                          |  62%  |                                                                              |====================================================                  |  75%  |                                                                              |=============================================================         |  88%  |                                                                              |======================================================================| 100%
#>   |Elapsed time: 0.6s
#>   |Simulate CERW with 10 steps
#>   |                                                                              |                                                                      |   0%  |                                                                              |=========                                                             |  12%  |                                                                              |==================                                                    |  25%  |                                                                              |==========================                                            |  38%  |                                                                              |===================================                                   |  50%  |                                                                              |============================================                          |  62%  |                                                                              |====================================================                  |  75%  |                                                                              |=============================================================         |  88%  |                                                                              |======================================================================| 100%
#>   |Elapsed time: 0.2s
#>          x       y        z         a        g            t            l
#> 1  2556476 1188336 1283.736 0.4477398 1.531790  0.068692653  0.008807867
#> 2  2559479 1189864 1348.457 0.4708216 1.551590  0.033344333  0.008807867
#> 3  2561040 1190510 1350.129 0.3924452 1.569807 -0.072700625  0.008807867
#> 4  2563527 1191694 1400.713 0.4441973 1.552436  0.068692653 -0.014467710
#> 5  2566234 1192916 1403.110 0.4239858 1.569989 -0.037352306  0.008807867
#> 6  2567543 1193582 1390.251 0.4707649 1.579552  0.033344333  0.008807867
#> 7  2569228 1194577 1373.728 0.5331988 1.579239  0.068692653  0.008807867
#> 8  2573480 1196964 1322.629 0.5115090 1.581276 -0.037352306  0.008807867
#> 9  2577669 1199244 1384.103 0.4986426 1.557908 -0.002003986 -0.014467710
#> 10 2581749 1201865 1513.318        NA       NA           NA           NA
#>           d            p
#> 1        NA           NA
#> 2  3347.312 2.346278e-18
#> 3  1774.123 1.914074e-18
#> 4  2673.088 4.172904e-19
#> 5  2897.830 8.049573e-18
#> 6  1547.246 4.392427e-18
#> 7  1998.864 5.417818e-18
#> 8  4920.502 5.797290e-18
#> 9  4695.761 7.097109e-18
#> 10 4850.981           NA