Unconditional Empirical Random Walk (UERW) in 3-D

This function creates unconditional walks with prescribed empirical properties (turning angle, lift angle and step length and the auto-differences of them. It can be used for uncon- ditional walks or to seed the conditional walks with comparably long simulations. The conditional walk connecting a given start with a certain end point by a given number of steps needs an attraction term (the Q probability, see qProb.3d) to ensure that the target is approached and hit. In order to calculate the Q probability for each step the distribution of turns and lifts to target and the distribution of distance to target has to be known. They can be derived from the empirical data (ideally), or estimated from an unconditional process with the same properties. Creates a unconditional empirical random walk, with a specific starting point, geometrically similar to the initial trajectory.

sim.uncond.3d(n.locs, start = c(0, 0, 0), a0, g0, densities, error = TRUE)

Arguments

n.locs

the number of locations for the simulated track

start

vector indicating the start point c(x,y,z)

a0

initial heading in radian

g0

initial gradient/polar angle in radian

densities

list object returned by the get.densities.3d function

error

logical: add random noise to the turn angle, lift angle and step length to account for errors measurements?

Value

A 3 dimensional trajectory in the form of a data.frame

Note

Simulations connecting start and end points with more steps than 1/10th or more of the number of steps of the empirical data should rather rely on simulated unconditional walks with the same properties than on the empirical data (factor = 1500).

Random initial heading

For a random initial heading a0 use: sample(atan2(diff(coordinates(track)[,2]), diff(coordinates(track)[,1])),1)

Examples

sim.uncond.3d(
  10, start = c(0, 0, 0), a0 = pi / 2, g0 = pi / 2,
  densities = get.track.densities.3d(niclas)
)
#>   |Simulate UERW with 10 steps
#>   |                                                                              |                                                                      |   0%
#>   |TLD cube dimensions: 9 x 14 x 3
#>   |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#>   |Elapsed time: 0s
#>             x         y          z        a        g            t            l
#> 1      0.0000     0.000    0.00000 1.570796 1.570796 -0.353684014 -0.052786249
#> 2    324.5782  2272.378  -82.16063 1.428920 1.606574 -0.141876685  0.035777672
#> 3  -1670.3136  7581.297 -257.12391 1.930235 1.601637  0.501315553 -0.004937077
#> 4  -3476.5945 12137.776 -478.03232 1.948213 1.615836  0.017977768  0.014198990
#> 5  -3682.0361 14809.111 -489.41739 1.647551 1.575046 -0.300661753 -0.040790219
#> 6  -3953.4209 20263.692 -468.90586 1.620509 1.567041 -0.027042315 -0.008005124
#> 7  -3206.4839 25423.717 -308.42798 1.427040 1.540027 -0.193468557 -0.027013937
#> 8  -1230.1563 31277.396 -161.76750 1.245191 1.547063 -0.181848908  0.007036163
#> 9    483.5555 36254.114 -301.96708 1.239168 1.597426 -0.006023873  0.050363373
#> 10  3762.7386 41630.778 -299.25750 1.023135 1.570366 -0.216032376 -0.027060088
#>           d            p
#> 1  5549.338           NA
#> 2  2296.912 7.257479e-05
#> 3  5674.048 5.835424e-05
#> 4  4906.420 1.483408e-04
#> 5  2679.247 7.430771e-05
#> 6  5461.366 1.918479e-04
#> 7  5216.276 3.112295e-04
#> 8  6180.043 3.983911e-04
#> 9  5265.377 2.912603e-04
#> 10 6297.743 4.385527e-04