Unconditional Empirical Random Walk (UERW) in 3-D

This function creates unconditional walks with prescribed empirical properties (turning angle, lift angle and step length and the auto-differences of them. It can be used for uncon- ditional walks or to seed the conditional walks with comparably long simulations. The conditional walk connecting a given start with a certain end point by a given number of steps needs an attraction term (the Q probability, see qProb.3d) to ensure that the target is approached and hit. In order to calculate the Q probability for each step the distribution of turns and lifts to target and the distribution of distance to target has to be known. They can be derived from the empirical data (ideally), or estimated from an unconditional process with the same properties. Creates a unconditional empirical random walk, with a specific starting point, geometrically similar to the initial trajectory.

sim.uncond.3d(n.locs, start = c(0, 0, 0), a0, g0, densities, error = TRUE)

Arguments

n.locs
the number of locations for the simulated track

start
vector indicating the start point c(x,y,z)

a0
initial heading in radian

g0
initial gradient/polar angle in radian

densities
list object returned by the get.densities.3d function

error
logical: add random noise to the turn angle, lift angle and step length to account for errors measurements?

Value

A 3 dimensional trajectory in the form of a data.frame

Note

Simulations connecting start and end points with more steps than 1/10th or more of the number of steps of the empirical data should rather rely on simulated unconditional walks with the same properties than on the empirical data (factor = 1500).

Random initial heading

For a random initial heading a0 use: sample(atan2(diff(coordinates(track)[,2]), diff(coordinates(track)[,1])),1)

Examples

sim.uncond.3d(10, start=c(0,0,0), a0=pi/2, g0=pi/2, densities=get.track.densities.3d(niclas))
#>   |Simulate UERW with 10 steps
#>   |                                                                              |                                                                      |   0%
#>   |TLD cube dimensions: 9 x 14 x 3
#>   |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%
#>   |Elapsed time: 0s
#>            x        y         z         a        g            t            l
#> 1     0.0000     0.00   0.00000 1.5707963 1.570796  0.170112206 -0.136125790
#> 2  -173.2031  5107.77  97.52172 1.6046931 1.551717  0.033896739 -0.019079533
#> 3  -339.4061 11525.86  73.03597 1.5966866 1.574610 -0.008006475  0.022893352
#> 4   693.0413 16724.68  76.12981 1.3747545 1.570213 -0.221932132 -0.004397523
#> 5  2319.4558 21819.17 148.72419 1.2617745 1.557223 -0.112979943 -0.012990051
#> 6  1015.4919 27201.05 213.55764 1.8085038 1.559089  0.546729297  0.001866397
#> 7  2260.9373 29111.39 113.10208 0.9930523 1.614818 -0.815451507  0.055729282
#> 8  2940.6483 31069.60 331.53603 1.2367009 1.465804  0.243648565 -0.149014094
#> 9  2320.2131 34853.41 269.67852 1.7333212 1.586927  0.496620282  0.121123286
#> 10 1626.8744 39782.87 203.35340 1.7105317 1.584119 -0.022789474 -0.002808207
#>           d            p
#> 1  1865.097           NA
#> 2  5111.636 7.819369e-05
#> 3  6420.283 3.346173e-04
#> 4  5300.351 4.449118e-04
#> 5  5348.308 4.162813e-04
#> 6  5537.965 8.862155e-05
#> 7  2282.681 2.875136e-05
#> 8  2084.306 1.713119e-05
#> 9  3834.837 1.385609e-04
#> 10 4978.424 1.753549e-04