Unconditional Empirical Random Walk (UERW) in 3-D

This function creates unconditional walks with prescribed empirical properties (turning angle, lift angle and step length and the auto-differences of them. It can be used for uncon- ditional walks or to seed the conditional walks with comparably long simulations. The conditional walk connecting a given start with a certain end point by a given number of steps needs an attraction term (the Q probability, see qProb.3d) to ensure that the target is approached and hit. In order to calculate the Q probability for each step the distribution of turns and lifts to target and the distribution of distance to target has to be known. They can be derived from the empirical data (ideally), or estimated from an unconditional process with the same properties. Creates a unconditional empirical random walk, with a specific starting point, geometrically similar to the initial trajectory.

sim.uncond.3d(n.locs, start = c(0, 0, 0), a0, g0, densities,
  error = TRUE)

Arguments

n.locs
the number of locations for the simulated track

start
vector indicating the start point c(x,y,z)

a0
initial heading in radian

g0
initial gradient/polar angle in radian

densities
list object returned by the get.densities.3d function

error
logical: add random noise to the turn angle, lift angle and step length to account for errors measurements?

Value

A 3 dimensional trajectory in the form of a data.frame

Note

Simulations connecting start and end points with more steps than 1/10th or more of the number of steps of the empirical data should rather rely on simulated unconditional walks with the same properties than on the empirical data (factor = 1500).

Random initial heading

For a random initial heading a0 use: sample(atan2(diff(coordinates(track)[,2]), diff(coordinates(track)[,1])),1)

Examples

sim.uncond.3d(10, start=c(0,0,0), a0=pi/2, g0=pi/2, densities=get.track.densities.3d(niclas))

#>   |Simulate UERW with 10 steps

#>   |                                                                              |                                                                      |   0%

#>   |TLD cube dimensions: 9 x 14 x 3

#>   |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#>   |Elapsed time: 0s

#>            x         y          z        a        g           t            l
#> 1     0.0000     0.000    0.00000 1.570796 1.570796  0.17011221  0.030553292
#> 2   673.4230  5302.229  -35.88874 1.444465 1.577511 -0.12633115  0.006714572
#> 3   934.9702  6884.340  -34.20519 1.406963 1.569746 -0.03750241 -0.007764436
#> 4  1504.5107 12460.294 -192.34744 1.469007 1.599004  0.06204432  0.029257052
#> 5  3187.3949 17427.519 -370.16612 1.244136 1.604689 -0.22487111  0.005685174
#> 6  5961.1470 22714.434 -511.38557 1.087628 1.594445 -0.15650815 -0.010243338
#> 7  6597.8452 24842.759 -494.76474 1.280115 1.563315  0.19248747 -0.031130621
#> 8  5956.6069 28037.603 -525.94983 1.768875 1.580366  0.48875954  0.017051509
#> 9  5714.1230 33570.229 -562.08738 1.614596 1.577322 -0.15427856 -0.003044553
#> 10 6375.1484 39646.580 -650.81303 1.462436 1.585311 -0.15216041  0.007989775
#>           d            p
#> 1  5549.338           NA
#> 2  5344.943 3.487730e-04
#> 3  1603.586 9.651060e-05
#> 4  5607.196 1.153426e-04
#> 5  5247.575 4.073857e-04
#> 6  5972.027 5.381414e-04
#> 7  2221.582 6.531913e-05
#> 8  3258.709 1.115678e-04
#> 9  5538.056 2.871586e-04
#> 10 6112.845 5.202505e-04