Unconditional Empirical Random Walk (UERW) in 3-D

This function creates unconditional walks with prescribed empirical properties (turning angle, lift angle and step length and the auto-differences of them. It can be used for uncon- ditional walks or to seed the conditional walks with comparably long simulations. The conditional walk connecting a given start with a certain end point by a given number of steps needs an attraction term (the Q probability, see qProb.3d) to ensure that the target is approached and hit. In order to calculate the Q probability for each step the distribution of turns and lifts to target and the distribution of distance to target has to be known. They can be derived from the empirical data (ideally), or estimated from an unconditional process with the same properties. Creates a unconditional empirical random walk, with a specific starting point, geometrically similar to the initial trajectory.

sim.uncond.3d(n.locs, start = c(0, 0, 0), a0, g0, densities, error = TRUE)

Arguments

n.locs
the number of locations for the simulated track

start
vector indicating the start point c(x,y,z)

a0
initial heading in radian

g0
initial gradient/polar angle in radian

densities
list object returned by the get.densities.3d function

error
logical: add random noise to the turn angle, lift angle and step length to account for errors measurements?

Value

A 3 dimensional trajectory in the form of a data.frame

Note

Simulations connecting start and end points with more steps than 1/10th or more of the number of steps of the empirical data should rather rely on simulated unconditional walks with the same properties than on the empirical data (factor = 1500).

Random initial heading

For a random initial heading a0 use: sample(atan2(diff(coordinates(track)[,2]), diff(coordinates(track)[,1])),1)

Examples

sim.uncond.3d(10, start=c(0,0,0), a0=pi/2, g0=pi/2, densities=get.track.densities.3d(niclas))

#>   |Simulate UERW with 10 steps

#>   |                                                                              |                                                                      |   0%

#>   |TLD cube dimensions: 9 x 14 x 3

#>   |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#>   |Elapsed time: 0s

#>             x         y        z         a        g           t             l
#> 1      0.0000     0.000   0.0000 1.5707963 1.570796 -0.35368401 -0.0527862491
#> 2   -118.8554  5265.499 102.4197 1.5933650 1.551353  0.02256865 -0.0194436895
#> 3   1218.4225 10298.675 378.4173 1.3111035 1.517849 -0.28226144 -0.0335037485
#> 4   2281.2192 11677.376 572.8942 0.9140725 1.459540 -0.39703100 -0.0583088029
#> 5   4960.1547 16258.470 635.6869 1.0416426 1.558965  0.12757009  0.0994245047
#> 6   7829.5692 20520.894 777.1202 0.9782923 1.543278 -0.06335037 -0.0156868005
#> 7  11863.7058 24742.304 883.7585 0.8080789 1.552535 -0.17021339  0.0092576248
#> 8  16737.6491 27988.124 983.8568 0.5875088 1.553704 -0.22057008  0.0011687835
#> 9  20774.3179 30332.151 830.1077 0.5260951 1.603722 -0.06141371  0.0500178437
#> 10 22532.7504 31938.748 749.5331 0.7403071 1.604612  0.21421206  0.0008899157
#>           d            p
#> 1  5549.338           NA
#> 2  5267.836 2.572652e-04
#> 3  5215.108 2.642547e-05
#> 4  1751.620 2.415129e-05
#> 5  5307.265 6.093709e-05
#> 6  5140.214 1.918079e-04
#> 7  5840.029 3.858087e-04
#> 8  5856.679 4.315515e-04
#> 9  4670.417 2.453195e-04
#> 10 2383.219 8.761572e-05