Transform coordinates reference system (CRS) of a 3-D track

Attention: Please use this function for CRS transformations, since it is based on the st_transform from the sf package and therefore supports CRS transformations in 3-D. Note: spTransform from the sp package only supports transformations in the 2D plane, which will cause distortions in the third dimension.

transformCRS.3d(track, fromCRS, toCRS)

Arguments

track
data.frame with x,y,z coordinates

fromCRS
string: proj4 of current CRS

toCRS
string: proj4 of CRS to be converted in

Value

A data.frame containing x,y,z and variables.

Examples

transformCRS.3d(niclas, fromCRS="+init=epsg:2056", toCRS="+init=epsg:4326")

#>           x        y       z           a            g           t             l
#> 1  6.868026 46.84473 1334.10  0.29022670 1.570782e+00          NA            NA
#> 2  6.895325 46.85352 1409.90  0.31124324 3.783110e-04  0.02101655 -1.570404e+00
#> 3  6.920342 46.85870 1420.00  0.20435409 2.529560e-03 -0.10688916  2.151249e-03
#> 4  6.945902 46.86607 1442.00  0.28087933 1.209199e-03  0.07652524 -1.320361e-03
#> 5  6.963975 46.86605 1856.80 -0.00148285 4.357117e-05 -0.28236218 -1.165628e-03
#> 6  7.017997 46.88671 1814.20  0.36526329 3.140235e+00  0.36674614  3.140191e+00
#> 7  7.082597 46.91257 1582.50  0.38074869 3.141292e+00  0.01548540  1.057370e-03
#> 8  7.137261 46.93549 1457.25  0.39705801 3.141119e+00  0.01630932 -1.729321e-04
#> 9  7.180323 46.95527 1332.00  0.43072179 3.141214e+00  0.03366378  9.487930e-05
#> 10 7.198807 46.96761 1563.10  0.58839577 9.615192e-05  0.15767398 -3.141118e+00
#> 11 7.230685 46.99044 1603.60  0.62150743 9.681450e-04  0.03311166  8.719930e-04
#> 12 7.268758 47.02529 1405.40  0.74131273 3.141332e+00  0.11980530  3.140364e+00
#> 13 7.306239 47.05830 1229.70  0.72196445 3.141308e+00 -0.01934828 -2.379902e-05
#> 14 7.324075 47.06770 1299.20  0.48533251 2.901432e-04 -0.23663194 -3.141018e+00
#> 15 7.337493 47.08205 1477.20  0.81887038 1.103582e-04  0.33353787 -1.797850e-04
#> 16 7.349967 47.08237 1497.50  0.02549749 6.146964e-04 -0.79337289  5.043382e-04
#> 17 7.401676 47.10885 1614.30  0.47335616 4.974034e-04  0.44785866 -1.172930e-04
#> 18 7.464232 47.12686 1535.80  0.28026606 3.140763e+00 -0.19309009  3.140266e+00
#> 19 7.525779 47.13273 1477.60  0.09502670 3.140530e+00 -0.18523937 -2.330293e-04
#> 20 7.596006 47.14952 1294.30  0.23476598 3.141199e+00  0.13973928  6.683512e-04
#> 21 7.629472 47.15956 1423.90  0.29125803 2.695750e-04  0.05649205 -3.140929e+00
#> 22 7.698594 47.17736 1348.40  0.25210505 3.140647e+00 -0.03915298  3.140378e+00
#> 23 7.742600 47.18721 1686.00  0.22007079 1.335701e-04 -0.03203426 -3.140514e+00
#> 24 7.817804 47.21143 1498.40  0.31162099 3.141171e+00  0.09155020  3.141038e+00
#> 25 7.889830 47.22471 1429.20  0.18231683 3.140534e+00 -0.12930416 -6.372317e-04
#> 26 7.971334 47.23750 1228.90  0.15567910 3.141181e+00 -0.02663773  6.464981e-04
#> 27 8.055092 47.22623 1328.60 -0.13377353 8.476821e-04 -0.28945263 -3.140333e+00
#> 28 8.131064 47.20756 1418.10 -0.24090985 8.740841e-04 -0.10713632  2.640200e-05
#> 29 8.188827 47.21677 1379.20  0.15812836 3.140089e+00  0.39903821  3.139215e+00
#> 30 8.218898 47.21298 1436.50 -0.12565518 5.289667e-04 -0.28378354 -3.139560e+00
#> 31 8.260129 47.22052 1136.80  0.18087906 3.141453e+00  0.30653424  3.140924e+00
#>            d
#> 1         NA
#> 2   75.80001
#> 3   10.10003
#> 4   22.00002
#> 5  414.80000
#> 6   42.60004
#> 7  231.70001
#> 8  125.25001
#> 9  125.25001
#> 10 231.10000
#> 11  40.50002
#> 12 198.20001
#> 13 175.70001
#> 14  69.50000
#> 15 178.00000
#> 16  20.30000
#> 17 116.80001
#> 18  78.50003
#> 19  58.20003
#> 20 183.30002
#> 21 129.60001
#> 22  75.50003
#> 23 337.60000
#> 24 187.60002
#> 25  69.20004
#> 26 200.30002
#> 27  99.70004
#> 28  89.50003
#> 29  38.90004
#> 30  57.30001
#> 31 299.70000